rocksdb: support memory views (#602)
* move createSession onto state

* need to special case draft state

* do not throw error if we are closing
