+++ lib/core.js
@@ -266,55 +266,6 @@ module.exports = class Core {
     this._manifestFlushed = !!header.manifest
   }
 
-  async _getTreeHeadAt (length) {
-    const head = getDefaultTree()
-
-    head.length = length
-
-    if (length === this.header.tree.length) {
-      head.fork = this.header.tree.fork
-      head.rootHash = this.header.tree.rootHash
-      return head
-    }
-
-    const roots = await this.tree.getRoots(length)
-    const rootHash = crypto.tree(roots)
-
-    head.fork = this.header.tree.fork
-    head.rootHash = rootHash
-
-    return head
-  }
-
-  async createSession (name, length, overwrite) {
-    let storage = null
-    let treeInfo = null
-
-    if (!overwrite) {
-      storage = await this.storage.openBatch(name)
-
-      if (storage !== null) {
-        treeInfo = (await getCoreHead(storage)) || getDefaultTree()
-        if (length !== -1 && treeInfo.length !== length) throw STORAGE_CONFLICT('Different batch stored here')
-      }
-    }
-
-    if (storage === null) {
-      treeInfo = await this._getTreeHeadAt(length === -1 ? this.tree.length : length)
-      storage = await this.storage.registerBatch(name, treeInfo)
-    }
-
-    const bitfield = await Bitfield.open(storage, treeInfo.length)
-
-    bitfield.merge(this.bitfield, treeInfo.length)
-
-    const tree = await MerkleTree.open(storage, treeInfo.length, {
-      prologue: this.tree.prologue
-    })
-
-    return new SessionState(this, storage, this.blocks, tree, -1, name)
-  }
-
   async audit () {
     await this.state.mutex.lock()
 
@@ -846,13 +797,6 @@ function parseHeader (info) {
 }
 
 function noop () {}
-
-function getCoreHead (storage) {
-  const b = storage.createReadBatch()
-  const p = b.getCoreHead()
-  b.tryFlush()
-  return p
-}
 
 async function getCoreInfo (storage) {
   const r = storage.createReadBatch()
