+++ lib/session-state.js
@@ -3,9 +3,12 @@ const b4a = require('b4a')
 const assert = require('nanoassert')
 const flat = require('flat-tree')
 const quickbit = require('quickbit-universal')
+
+const { STORAGE_CONFLICT, INVALID_OPERATION, INVALID_SIGNATURE } = require('hypercore-errors')
+
 const Mutex = require('./mutex')
 const Bitfield = require('./bitfield')
-const { INVALID_OPERATION, INVALID_SIGNATURE } = require('hypercore-errors')
+const MerkleTree = require('./merkle-tree')
 
 module.exports = class SessionState {
   constructor (core, storage, blocks, tree, snapshotLength, name) {
@@ -525,6 +528,50 @@ module.exports = class SessionState {
       this.mutex.unlock()
     }
   }
+
+  async _getTreeHeadAt (length) {
+    const head = getDefaultTree()
+
+    head.length = length
+
+    const roots = await this.tree.getRoots(length)
+    const rootHash = crypto.tree(roots)
+
+    head.fork = this.tree.fork
+    head.rootHash = rootHash
+
+    return head
+  }
+
+  async createSession (name, length, overwrite, draft) {
+    let storage = null
+    let treeInfo = null
+
+    if (!overwrite && this.storage) {
+      storage = await this.storage.openBatch(name)
+
+      if (storage !== null) {
+        treeInfo = (await getCoreHead(storage)) || getDefaultTree()
+        if (length !== -1 && treeInfo.length !== length) throw STORAGE_CONFLICT('Different batch stored here')
+      }
+    }
+
+    if (storage === null) {
+      treeInfo = await this._getTreeHeadAt(length === -1 ? this.tree.length : length)
+
+      if (draft !== true) {
+        storage = await this.storage.registerBatch(name, treeInfo)
+      } else {
+        storage = await this.storage.registerOverlay(treeInfo)
+      }
+    }
+
+    const tree = await MerkleTree.open(storage, treeInfo.length, {
+      prologue: this.tree.prologue
+    })
+
+    return new SessionState(this.core, storage, this.core.blocks, tree, -1, name)
+  }
 }
 
 function noop () {}
@@ -589,4 +636,20 @@ function updateDependency (state, length) {
     data: target.data,
     length
   }
+}
+
+function getDefaultTree () {
+  return {
+    fork: 0,
+    length: 0,
+    rootHash: null,
+    signature: null
+  }
+}
+
+function getCoreHead (storage) {
+  const b = storage.createReadBatch()
+  const p = b.getCoreHead()
+  b.tryFlush()
+  return p
 }
